Add first and simple person implementation
A person has a first name, a surname, a display name, an email address, and also a username and
password.

At the very minimum, only a first name, a surname and a username are needed (username's purpose is
to identify a person, instead of a number).

A simple comparison is provided : two persons compare as equal if they have identical first names,
surnames, usernames and email addresses, that is: display names and passwords are ignored.